+++ src/biz/bokhorst/xprivacy/Meta.java
@@ -227,6 +227,9 @@ public class Meta {
 		mListHook.add(new Hook("location", "GMS.requestLocationUpdates", "ACCESS_COARSE_LOCATION,ACCESS_FINE_LOCATION", 1, null, null).unsafe().optional());
 		mListHook.add(new Hook("location", "GMS.requestActivityUpdates", "com.google.android.gms.permission.ACTIVITY_RECOGNITION", 1, null, null).unsafe());
 
+		mListHook.add(new Hook("location", "GMS5.getLastLocation", "ACCESS_COARSE_LOCATION,ACCESS_FINE_LOCATION", 1, "2.99.26", null).unsafe().optional());
+		mListHook.add(new Hook("location", "GMS5.requestLocationUpdates", "ACCESS_COARSE_LOCATION,ACCESS_FINE_LOCATION", 1, "2.99.26", null).unsafe().optional());
+
 		mListHook.add(new Hook("location", "MapV1.enableMyLocation", "ACCESS_COARSE_LOCATION,ACCESS_FINE_LOCATION", 1, "2.1.25", null).unsafe().optional());
 
 		mListHook.add(new Hook("location", "MapV2.getMyLocation", "ACCESS_COARSE_LOCATION,ACCESS_FINE_LOCATION", 1, "2.1.25", null).unsafe().optional());
@@ -529,8 +532,8 @@ public class Meta {
 		mListHook.add(new Hook(null, "getSystemService", "", 1, null, null).notAOSP(19));
 
 		// GoogleApiClient.Builder
-		mListHook.add(new Hook(null, "GAC.addConnectionCallbacks", "", 1, null, null).optional());
-		mListHook.add(new Hook(null, "GAC.onConnected", "", 1, null, null));
+		mListHook.add(new Hook(null, "GMS5.addConnectionCallbacks", "", 1, null, null).optional());
+		mListHook.add(new Hook(null, "GMS5.onConnected", "", 1, null, null));
 
 		// IntentFirewall
 		mListHook.add(new Hook(null, "checkIntent", "", 19, null, null));
@@ -539,6 +542,9 @@ public class Meta {
 		mListHook.add(new Hook(null, "GMS.removeActivityUpdates", "", 1, null, null));
 		mListHook.add(new Hook(null, "GMS.removeGeofences", "", 1, null, null).optional());
 		mListHook.add(new Hook(null, "GMS.removeLocationUpdates", "", 1, null, null).optional());
+
+		// FusedLocationProviderApi
+		mListHook.add(new Hook(null, "GMS5.removeLocationUpdates", "", 1, "2.99.26", null).optional());
 
 		// LocationManager/Service
 		mListHook.add(new Hook(null, "removeUpdates", "", 3, null, null).notAOSP(19));