+++ src/biz/bokhorst/xprivacy/XGoogleApiClient.java
@@ -3,11 +3,14 @@ package biz.bokhorst.xprivacy;
 import java.util.ArrayList;
 import java.util.List;
 
+import android.os.Binder;
+import android.util.Log;
+
 public class XGoogleApiClient extends XHook {
 	private Methods mMethod;
 
 	private XGoogleApiClient(Methods method, String restrictionName) {
-		super(restrictionName, method.name(), "GAC." + method.name());
+		super(restrictionName, method.name(), "GMS5." + method.name());
 		mMethod = method;
 	}
 
@@ -27,7 +30,7 @@ public class XGoogleApiClient extends XHook {
 	};
 
 	public static List<XHook> getInstances() {
-		Util.log(null, android.util.Log.WARN, "Loaded GAC");
+		Util.log(null, Log.INFO, "Loaded GoogleApiClient$Builder uid=" + Binder.getCallingUid());
 		List<XHook> listHook = new ArrayList<XHook>();
 		listHook.add(new XGoogleApiClient(Methods.addConnectionCallbacks, null));
 		return listHook;
@@ -41,6 +44,7 @@ public class XGoogleApiClient extends XHook {
 				Class<?> clazz = param.args[0].getClass();
 				if (PrivacyManager.getTransient(clazz.getName(), null) == null) {
 					PrivacyManager.setTransient(clazz.getName(), Boolean.toString(true));
+
 					XPrivacy.hookAll(XConnectionCallbacks.getInstances(param.args[0]), clazz.getClassLoader(),
 							getSecret());
 				}